Release/1.21.1

See merge request infopaginas/infopaginas!1650

+++ src/Domain/ReportBundle/Command/ReportCleanUpDataCommand.php
@@ -7,6 +7,8 @@ use Domain\SiteBundle\Logger\CronLogger;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Lock\Factory;
+use Symfony\Component\Lock\Store\SemaphoreStore;
 
 /**
  * Class ReportCleanUpDataCommand
@@ -17,12 +19,12 @@ class ReportCleanUpDataCommand extends ContainerAwareCommand
     /**
      * Lifetime of mongoDB logs
      */
-    const LOGS_LIFETIME = 5;
+    private const LOGS_LIFETIME = 5;
 
     protected function configure()
     {
         $this
-            ->setName('domain:report-mongo-db:clean-up')
+            ->setName(CronLogger::MONGO_CLEAN_UP)
             ->setDescription('Clean up mongoDB reports')
         ;
     }
@@ -33,28 +35,40 @@ class ReportCleanUpDataCommand extends ContainerAwareCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $container = $this->getContainer();
-        $logger = $container->get('domain_site.cron.logger');
-        $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_START, CronLogger::MESSAGE_START);
+        $store = new SemaphoreStore();
+        $factory = new Factory($store);
+        $lock = $factory->createLock(CronLogger::MONGO_CLEAN_UP);
 
-        $rawDataArchivingDate = $this->getRawDataDeletionDate();
+        if ($lock->acquire()) {
+            $container = $this->getContainer();
+            $logger = $container->get('domain_site.cron.logger');
+            $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_START, CronLogger::MESSAGE_START);
 
-        $output->writeln('Start...');
+            $rawDataArchivingDate = $this->getRawDataDeletionDate();
 
-        $output->writeln('Process overview report');
-        $this->getBusinessReportManager()->deleteArchivedRawBusinessInteractions($rawDataArchivingDate);
-        $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process overview report');
+            $output->writeln('Start...');
 
-        $output->writeln('Process keyword report');
-        $this->getKeywordReportManager()->deleteArchivedRawKeywordData($rawDataArchivingDate);
-        $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process keyword report');
+            $output->writeln('Process overview report');
+            $this->getBusinessReportManager()->deleteArchivedRawBusinessInteractions($rawDataArchivingDate);
+            $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process overview report');
 
-        $output->writeln('Process category report');
-        $this->getCategoryReportManager()->deleteArchivedRawCategoryData($rawDataArchivingDate);
-        $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process category report');
+            $output->writeln('Process keyword report');
+            $this->getKeywordReportManager()->deleteArchivedRawKeywordData($rawDataArchivingDate);
+            $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process keyword report');
 
-        $output->writeln('done');
-        $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_END, CronLogger::MESSAGE_STOP);
+            $output->writeln('Process category report');
+            $this->getCategoryReportManager()->deleteArchivedRawCategoryData($rawDataArchivingDate);
+            $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_IN_PROGRESS, 'execute:Process category report');
+
+            $output->writeln('done');
+            $logger->addInfo(CronLogger::MONGO_CLEAN_UP, CronLogger::STATUS_END, CronLogger::MESSAGE_STOP);
+
+            $lock->release();
+        } else {
+            return $output->writeln('Command is locked by another process');
+        }
+
+        return 0;
     }
 
     /**
